ARCHIVE/old_configs/CompassBuilder_Starter_Kit/requirements.txt
python-docx
PyPDF2
PyYAML
